feat: refactored unicorn web into separate modules - ApprovalService and SearchService - to align with implementations in other runtimes.
