+++ unicorn_web/ApprovalService/src/main/java/schema/unicorn_properties/publicationevaluationcompleted/AWSEvent.java
