+++ unicorn_web/ApprovalService/src/main/resources/log4j2.xml
@@ -9,6 +9,6 @@
         <Root level="WARN">
             <AppenderRef ref="ConsoleAppender"/>
         </Root>
-        <Logger name="property" level="DEBUG" />
+        <Logger name="search" level="DEBUG" />
     </Loggers>
 </Configuration>
