+++ unicorn_web/SearchService/tests/events/eventbridge/put_event_publication_evaluation_completed.json
@@ -1,0 +1,8 @@
+[
+  {
+    "Source": "unicorn.properties",
+    "Detail": "{\"property_id\":\"usa/anytown/main-street/111\",\"evaluation_result\": \"APPROVED\"}",
+    "DetailType": "PublicationEvaluationCompleted",
+    "EventBusName": "UnicornWebBus-local"
+  }
+]
