+++ unicorn_web/integration/event-schemas.yaml
@@ -53,7 +53,7 @@ Resources:
         Fn::GetAtt: EventRegistry.RegistryName
       SchemaName:
         Fn::Sub: '{{resolve:ssm:/uni-prop/UnicornWebNamespace}}@PublicationApprovalRequested'
-      Description: 'The schema for a request to publish a property'
+      Description: 'The schema for a request to publish a search'
       Content:
         Fn::Sub: |
           {
